backend/uploader_service/requirements.txt
fastapi
urllib3
uvicorn
boto3
pillow
python-dotenv
jinja2
python-multipart